Organize Library Into Package
Made the library into its own package nn. This allows for better organization and easier import functionality. Cuda support has been removed. Modified all training examples and README for the updated import.
