--- config.py
@@ -1,2 +1,0 @@
-#Global Toggle For CUDA speedup. Only works if you have CUDA Toolkit installed
-enableCuda = False
