+++ nn/data_processing/file_io.py
@@ -1,5 +1,5 @@
 import dill
-from network import Network
+from nn import Network
 
 #Provides the ability to save and load neural network objects from files.
 #The entire network is saved as an object to a file.
