+++ nn/layers/dense.py
@@ -1,9 +1,5 @@
-from config import *
-if enableCuda:
-    import cupy as np
-else:
-    import numpy as np
-from layer import Layer
+import numpy as np
+from nn.interfaces.layer import Layer
 
 class Dense(Layer):
     def __init__(self, input_size, output_size):
