+++ requirements.txt
@@ -3,4 +3,3 @@ matplotlib==3.5.1
 tensorflow==2.7.0
 scipy==1.7.1
 dill==0.3.5.1
-cupy-cuda11x
